+++ src/Pickles/Pickles.TestFrameworks/MsTest/MsTestResults.cs
@@ -25,7 +25,7 @@ namespace PicklesDoc.Pickles.TestFrameworks.MsTest
     public class MsTestResults : MultipleTestRunsBase
     {
         public MsTestResults(IConfiguration configuration, MsTestSingleResultLoader singleResultLoader, MsTestExampleSignatureBuilder exampleSignatureBuilder)
-            : base(false, configuration, singleResultLoader, exampleSignatureBuilder)
+            : base(true, configuration, singleResultLoader, exampleSignatureBuilder)
         {
         }
     }